client location pickup functions

+++ app/src/main/java/ua/com/programmer/agentventa/catalogs/locations/pickup/LocationPickupFragment.kt
@@ -1,6 +1,7 @@
 package ua.com.programmer.agentventa.catalogs.locations.pickup
 
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.Menu
 import android.view.MenuInflater
@@ -20,6 +21,7 @@ import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.GoogleMap.OnMarkerDragListener
 import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.BitmapDescriptorFactory
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.Marker
 import com.google.android.gms.maps.model.MarkerOptions
@@ -27,6 +29,7 @@ import com.google.maps.android.collections.MarkerManager
 import dagger.hilt.android.AndroidEntryPoint
 import ua.com.programmer.agentventa.R
 import ua.com.programmer.agentventa.dao.entity.ClientLocation
+import ua.com.programmer.agentventa.dao.entity.LocationHistory
 import ua.com.programmer.agentventa.dao.entity.hasLocation
 import ua.com.programmer.agentventa.databinding.LocationPickupFragmentBinding
 import ua.com.programmer.agentventa.extensions.format
@@ -43,8 +46,11 @@ class LocationPickupFragment: Fragment(), MenuProvider, OnMapReadyCallback {
     private val binding get() = _binding!!
 
     private var clientLocation: ClientLocation? = null
+    private var currentLocation: LocationHistory? = null
     private var map: GoogleMap? = null
     private var markerCollection: MarkerManager.Collection? = null
+
+    private val markerTagCurrentLocation = "current_location"
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -66,6 +72,10 @@ class LocationPickupFragment: Fragment(), MenuProvider, OnMapReadyCallback {
         // account is necessary to save new location
         sharedViewModel.currentAccount.observe(viewLifecycleOwner) {
             viewModel.setAccountGuid(it.guid)
+        }
+        viewModel.currentLocation.observe(viewLifecycleOwner) {
+            currentLocation = it
+            showCurrentLocation(it)
         }
         viewModel.clientLocation.observe(viewLifecycleOwner) {
             clientLocation = it
@@ -105,10 +115,17 @@ class LocationPickupFragment: Fragment(), MenuProvider, OnMapReadyCallback {
                 showLocation()
             }
             R.id.clear_location -> {
-                if (viewModel.onDeleteLocation()) {
-                    markerCollection?.clear()
-                    Toast.makeText(requireContext(), R.string.data_deleted, Toast.LENGTH_SHORT).show()
-                }
+                AlertDialog.Builder(requireContext())
+                    .setTitle(R.string.warning)
+                    .setMessage(R.string.text_erase_data)
+                    .setPositiveButton(R.string.OK) { _, _ ->
+                        if (viewModel.onDeleteLocation()) {
+                            deleteClientMarker()
+                            Toast.makeText(requireContext(), R.string.data_deleted, Toast.LENGTH_SHORT).show()
+                        }
+                    }
+                    .setNegativeButton(R.string.cancel) { _, _ -> }
+                    .show()
             }
             else -> return false
         }
@@ -135,16 +152,70 @@ class LocationPickupFragment: Fragment(), MenuProvider, OnMapReadyCallback {
                         }
                 }
             )
+            markerCollection?.setOnMarkerClickListener {
+                Log.d("LocationPickupFragment", "marker: $it")
+                if (it.tag == markerTagCurrentLocation && viewModel.canEditLocation) {
+                    AlertDialog.Builder(requireContext())
+                        .setTitle(R.string.warning)
+                        .setMessage(R.string.use_current_location)
+                        .setPositiveButton(R.string.OK) { _, _ ->
+                            viewModel.useCurrentLocation()
+                        }
+                        .setNegativeButton(R.string.cancel) { _, _ -> }
+                        .show()
+                }
+                true
+            }
             viewModel.setClientParameters(navigationArgs.clientGuid)
         }
+    }
 
+    private fun currentLocationOptions(location: LocationHistory): MarkerOptions {
+        val latLng = LatLng(location.latitude, location.longitude)
+        return MarkerOptions()
+            .position(latLng)
+            //.flat(true)
+            //.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_AZURE))
+            .icon(BitmapDescriptorFactory.fromResource(R.drawable.map_marker_car_48))
+            .anchor(0.5f, 1f)
+    }
+
+    private fun clientLocationOptions(location: ClientLocation): MarkerOptions {
+        val latLng = LatLng(location.latitude, location.longitude)
+        return MarkerOptions()
+            .position(latLng)
+            .draggable(viewModel.canEditLocation)
+            .icon(BitmapDescriptorFactory.fromResource(R.drawable.map_marker_pin_red_48))
+            .anchor(0.5f, 1f)
+    }
+
+    private fun showCurrentLocation(location: LocationHistory?) {
+        if (location == null || clientLocation?.hasLocation() == true) return
+        markerCollection?.let { m ->
+            val latLng = LatLng(location.latitude, location.longitude)
+            val marker = m.markers?.find {
+                it.tag == markerTagCurrentLocation
+            } ?: m.addMarker(
+                currentLocationOptions(location)
+            )
+            marker?.apply {
+                tag = markerTagCurrentLocation
+                position = latLng
+                //rotation = location.bearing.toFloat()
+            }
+        }
     }
 
     private fun showLocation() {
-        markerCollection?.clear()
-        //binding.bottomBar.visibility = View.VISIBLE
         val hasLocation = clientLocation?.hasLocation() ?: false
         if (!hasLocation) {
+            currentLocation?.let {
+                val latLng = LatLng(it.latitude, it.longitude)
+                map?.moveCamera(com.google.android.gms.maps.CameraUpdateFactory.newLatLngZoom(
+                    latLng, 17f
+                ))
+                showCurrentLocation(it)
+            }
             val alert = AlertDialog.Builder(requireContext())
             alert.setTitle(R.string.warning)
             alert.setMessage(R.string.warn_no_location)
@@ -154,21 +225,31 @@ class LocationPickupFragment: Fragment(), MenuProvider, OnMapReadyCallback {
         }
         map?.let { m ->
             clientLocation?.let { c ->
-                val location = LatLng(c.latitude, c.longitude)
-                val options = MarkerOptions()
-                    .position(location)
-                    .draggable(viewModel.canEditLocation)
-                val marker = markerCollection?.addMarker(options)
+                val marker = markerCollection?.markers?.find {
+                    it.tag == c.clientGuid
+                } ?: markerCollection?.addMarker(clientLocationOptions(c))
                 marker?.apply {
                     tag = c.clientGuid
                     //title = c.description
                     snippet = c.address
                 }
+                val location = LatLng(c.latitude, c.longitude)
                 binding.coordinates.text = locationText(location)
                 m.moveCamera(com.google.android.gms.maps.CameraUpdateFactory.newLatLngZoom(
                         location, 17f
                     ))
+                markerCollection?.markers?.find {
+                    it.tag == markerTagCurrentLocation
+                }?.remove()
             }
+        }
+    }
+
+    private fun deleteClientMarker() {
+        clientLocation?.let { c ->
+            markerCollection?.markers?.find {
+                it.tag == c.clientGuid
+            }?.remove()
         }
     }
